Add files via upload
An interactive and visually appealing landing page designed for a restaurant, utilizing HTML, CSS, and JavaScript. This project showcases a seamless user experience with a stunning layout, dynamic features, and a responsive design that captures the essence of [Restaurant Name]. Perfect for attracting food enthusiasts and providing essential information at a glance.